Relocate verifier event-emit + repair terminal helpers (batch 10)
Phase 2 batch 10 for Issue #682. Moves 6 verifier-progress /
patch-proposal event-emit helpers + 1 struct (~183 LOC of bodies)
out of turn.rs into verifier_orchestration.rs.

Moved (turn.rs -> verifier_orchestration.rs):

  Event emitters (3):
    - emit_repair_progress_classified_event (38 LOC) — emits
      `agent.verifier_repair.progress_classified` event using
      super::repair_progress::classify_repair_progress
    - emit_patch_proposal_shadow_validation_event (56 LOC) — emits
      `agent.verifier_patch_proposal.shadow_validation` event +
      runs the patch-provider admission shadow validator
    - emit_patch_proposal_legacy_validation_comparison_event (39 LOC) —
      emits `agent.verifier_patch_proposal.validation_comparison`
      event with the shadow-vs-legacy decisive-agreement signal

  Shadow validator support:
    - patch_proposal_target_contents_for_shadow (18 LOC) — symlink-
      containment-safe file reader for the shadow validator
    - PatchProposalShadowValidation struct + impl (status / reason /
      is_decisive / accepted) used by both shadow + legacy comparison
      events

  Terminal exit reason mapping:
    - repair_terminal_exit_reason (7 LOC) — maps RepairTerminalReason
      to ExitReason for the repair-driver dispatch

All 6 fns + struct promoted to pub(super). Struct fields + impl
methods are also pub(super) so call sites can construct literals.

Adjustments:

  * verifier_orchestration.rs: added imports for
    - super::patch_proposal::PatchProposal
    - super::patch_provider::{PatchProviderKind, PatchProviderOutput,
      PatchProviderRequest, admit_patch_provider_output}
    - super::repair_job::{RepairTerminalReason, StopReason}
    - super::repair_progress::classify_repair_progress
    - super::summary::ExitReason
    - crate::logging::{log_llm_event, stable_path_hash}

  * turn.rs:
    - Extended use super::verifier_orchestration::{...} with the
      6 production-callable relocated fns.
    - Added `use super::super::verifier_orchestration::PatchProposalShadowValidation;`
      to the `mod repair_lifecycle_event_tests` block so the
      `PatchProposalShadowValidation { ... }` literal in the
      `patch_shadow_old_string_not_found_defers_to_production_validator`
      test resolves (the existing `use super::*;` no longer reaches
      the struct now that it has moved out of turn.rs).
    - Dropped `VERIFIER_REPAIR_PASS_MAX_FILE_BYTES` from the
      `super::repair_driver` re-import (only the relocated
      `patch_proposal_target_contents_for_shadow` used it).

Metrics: turn.rs 33,333 -> 33,149 (-184);
verifier_orchestration.rs 1,724 -> 1,920 (+196). Cumulative Issue
#682 progress: turn.rs 34,958 -> 33,149 (-1,809, ~60% of Phase 2
target). Acceptance gap: 33,149 -> 32,000 = ~1,149 LOC remaining.
